requirements.txt
PyYAML
WeasyPrint
jinja2
